Requested changes part 2
