+++ .gitignore
@@ -1,3 +1,4 @@
+.DS_Store
 /.build
 /Packages
 xcuserdata
